file_mover/requirements.txt
aiofiles==24.1.0
psutil==7.0.0
python_pushover2==0.5
Requests==2.32.3
